Officer Bardtolino/requirements.txt
aiogram==2.25.1
mysql-connector-python==8.3.0
captcha==0.5.0
python-dotenv==1.0.1